Source/requirements.txt
matplotlib>=3.4.3
twitch-python==0.0.19
twitchio==2.0.6
